GeobukSchool/school-record-parser-server-main/school-record-parser-server-main/requirements.txt
fastapi
uvicorn
python-multipart
PyPDF2
langchain
langchain-community
langchain-openai
openai
faiss-cpu
python-dotenv
pydantic-settings
python-magic
slowapi